Изменения в работе Favourites и Tray
- Убрана проверка File.Exists для Favourites.FilePath (поддержка URL)
- Добавлено подтверждение при удалении из Favourites
- Перенесена кнопка открытия Favourites в Tray
- Реализовано MVVM для Tray меню Favourites
- Добавлен пункт "Перезагрузить все службы" в Tray

+++ MyServer/UserControls/FavouriteUserControl.xaml.cs
@@ -29,7 +29,11 @@ namespace MyServer.UserControls
         {
             if (_viewModel.SelectedFavourite != null && _viewModel.SelectedFavourite is Favourite)
             {
-                FavouriteStore.Instance.DeleteFavourite(_viewModel.SelectedFavourite.Id);
+                MessageBoxResult result = MessageBox.Show("Вы уверены, что хотите удалить избранное ?", "Удаление избранного", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes)
+                {
+                    FavouriteStore.Instance.DeleteFavourite(_viewModel.SelectedFavourite.Id);
+                }
             }
             else
             {
